feat: fix PR comments
